+++ python/hsfs/core/hudi_engine.py
@@ -235,6 +235,20 @@ class HudiEngine:
 
         return hudi_options
 
+    def reconcile_hudi_schema(self, spark_engine, hudi_fg_alias, read_options):
+        fg_table_name = hudi_fg_alias.feature_group._get_table_name()
+        if (
+                self._spark_session.table(hudi_fg_alias.alias).schema
+                != self._spark_session.table(fg_table_name).schema
+        ):
+            dataframe = self._spark_session.table(fg_table_name).limit(0)
+            spark_engine.save_empty_dataframe(hudi_fg_alias.feature_group, dataframe)
+
+            self.register_temporary_table(
+                hudi_fg_alias,
+                read_options,
+            )
+
     @staticmethod
     def _get_last_commit_metadata(spark_context, base_path):
         hopsfs_conf = spark_context._jvm.org.apache.hadoop.fs.FileSystem.get(
